Refactoring the menu function
変更箇所
・handlers/web/menu.go
HTTPリクエストの処理とビューのレンダリングの役割のみに修正
・handlers/api/menu.go
判定処理のロジック部分をまとめた
・public/json/food_stores.json
データの追加
・menu.html
formの送信先の修正
