+++ public/json/food_stores.json
@@ -1,0 +1,72 @@
+[
+  {
+    "name": "メキシカンフード",
+    "genre": "メキシカンフード",
+    "url": "https://www.toranomonhills.com/gourmet_shops/7537.html"
+  },
+  {
+    "name": "鮎ラーメン +",
+    "genre": "ラーメン",
+    "url": "https://www.toranomonhills.com/toranomonyokocho/1022.html"
+  },
+  {
+    "name": "AM STRAM GRAM",
+    "genre": "タルト",
+    "url": "https://www.toranomonhills.com/gourmet_shops/3617.html"
+  },
+  {
+    "name": "意気な寿司処阿部",
+    "genre": "寿司",
+    "url": "https://www.toranomonhills.com/gourmet_shops/0008.html"
+  },
+  {
+    "name": "餃子マニア",
+    "genre": "中華",
+    "url": "https://www.toranomonhills.com/gourmet_shops/3642.html"
+  },
+  {
+    "name": "CARAVAN",
+    "genre": "ピザ",
+    "url": "https://www.toranomonhills.com/gourmet_shops/7077.html"
+  },
+  {
+    "name": "Cassolo",
+    "genre": "イタリアン",
+    "url": "https://www.toranomonhills.com/gourmet_shops/0096.html"
+  },
+  {
+    "name": "焼千房 虎ノ門",
+    "genre": "鉄板焼き",
+    "url": "https://www.toranomonhills.com/gourmet_shops/0025.html"
+  },
+  {
+    "name": "スタバ",
+    "genre": "カフェ",
+    "url": "https://www.toranomonhills.com/gourmet_shops/0022.html"
+  },
+  {
+    "name": "スコインター",
+    "genre": "タイ料理",
+    "url": "https://www.toranomonhills.com/gourmet_shops/5575.html"
+  },
+  {
+    "name": "鰻まえはら",
+    "genre": "鰻",
+    "url": "https://www.toranomonhills.com/gourmet_shops/0027.html"
+  },
+  {
+    "name": "京菓匠 鶴屋吉信",
+    "genre": "和菓子",
+    "url": "https://www.toranomonhills.com/gourmet_shops/0041.html"
+  },
+  {
+    "name": "BeBu(ビブ)",
+    "genre": "ハンバーガー",
+    "url": "https://www.toranomonhills.com/gourmet_shops/0011.html"
+  },
+  {
+    "name": "なるたけ虎ノ門",
+    "genre": "和食",
+    "url": "https://www.toranomonhills.com/gourmet_shops/0093.html"
+  }
+]
